requirements.txt
fastapi
uvicorn
python-dotenv
configparser
bcrypt
pyjwt
starlette
python-multipart
pydantic
pydantic[email]
pymongo
httpx
pre-commit
